Corregir definición de las calificaciones

+++ db/crear_db.sql
@@ -17,14 +17,14 @@ DROP TABLE IF EXISTS usuarios;
 CREATE TABLE calificaciones_producto (
 	id_producto BIGINT(20) NOT NULL,
 	id_usuario BIGINT(20) NOT NULL,
-	calificacion TINYINT(1) NOT NULL,
+	calificacion DECIMAL(2,1) NOT NULL,
 	descripcion VARCHAR(255),
 	PRIMARY KEY (id_producto, id_usuario)
 );
 CREATE TABLE calificaciones_tienda (
 	id_tienda BIGINT(20) NOT NULL,
 	id_usuario BIGINT(20) NOT NULL,
-	calificacion TINYINT(1) NOT NULL,
+	calificacion DECIMAL(2,1) NOT NULL,
 	descripcion VARCHAR(255),
 	PRIMARY KEY (id_usuario, id_tienda)
 );
